Add features spec to create seasons
